améliorations authentification renforcée

+++ backend/controllers/user.js
@@ -6,14 +6,16 @@ const cryptoJS = require('crypto-js');
 const jwt = require('jsonwebtoken');
 
 exports.signup = (req, res, next) => {
+    //hashe le mot de passe
     bcrypt.hash(req.body.password, 10)
         .then(hash => {
             //définie les paramètres pour le cryptage de l'email
             const key = cryptoJS.enc.Hex.parse(process.env.CRYPTO_KEY);
             const iv = cryptoJS.enc.Hex.parse(process.env.CRYPTO_IV);
 
+            //crée une nouvelle instance de User
             User.create({
-                email: cryptoJS.AES.encrypt(req.body.email, key, {iv: iv}).toString(),
+                email: cryptoJS.AES.encrypt(req.body.email, key, {iv: iv}).toString(), //crypte l'email
                 password: hash,
                 firstName: req.body.firstName,
                 lastName: req.body.lastName
@@ -25,15 +27,17 @@ exports.signup = (req, res, next) => {
 };
 
 exports.login = (req, res, next) => {
+    //cherche le user qui correspond à l'email
     User.findOne({ where: { email: req.body.email } })
         .then(user => {
-            if (!user) {
+            if (!user) { //si le user n'existe pas
                 return res.status(401).json({ error: 'Utilisateur non trouvé !' });
             }
 
+            //compare le hash des mots de passe
             bcrypt.compare(req.body.password, user.password)
                 .then(valid => {
-                    if (!valid) {
+                    if (!valid) {  //si le mot de passe est invalide
                         return res.status(401).json({ error: 'Mot de passe invalide !' });
                     }
                     res.status(200).json({
@@ -41,7 +45,7 @@ exports.login = (req, res, next) => {
                         token: jwt.sign(
                             { userId: user.id },
                             process.env.TOKEN_KEY,
-                            { expiresIn: '2h'}
+                            { expiresIn: '24h'}
                         )
                     });
                 })
@@ -51,9 +55,20 @@ exports.login = (req, res, next) => {
 };
 
 exports.deleteAccount = (req, res, next) => {
-    User.destroy({ where: { id: req.params.id }} )
-        .then(() => res.status(200).json({ message: 'Utilisateur supprimé !' }))
+    //cherche le user qui correspond à l'id en paramètre
+    User.findOne({ where: {id: req.params.id} })
+        .then(oneUser => {
+            //vérifie si l'id correspond à l'id du token stocké dans req.id ou si l'utilisateur est administrateur
+            if (oneUser.id !== req.user.id) {
+                return res.status(403).json({ error: "Seul l'utilisateur peut supprimer son compte !"})
+            } else {
+                User.destroy({ where: { id: req.params.id }} )
+                    .then(() => res.status(200).json({ message: 'Utilisateur supprimé !' }))
+                    .catch(error => res.status(400).json({ error }));
+            }
+        })
         .catch(error => res.status(400).json({ error }));
+
 };
 
 exports.getUser = (req, res, next) => {
